update(*): divide code by modules
